Fix BaseViewModel import and type conflicts, implement camera methods, integrate GSR recorder with UI, and connect settings functionality

+++ app/src/main/java/mpdc4gsr/feature/thermal/data/source/TopdonDataSourceImpl.kt
@@ -272,7 +272,7 @@ class TopdonDataSourceImpl(
 
             LibIRProcess.convertYuyvMapToARGBPseudocolor(
                 frame,
-                CAMERA_WIDTH * CAMERA_HEIGHT,
+                (CAMERA_WIDTH * CAMERA_HEIGHT).toLong(),
                 CommonParams.PseudoColorType.PSEUDO_1,
                 rgbBuffer
             )
